Rename Test classes with the mandatory s suffix
