+++ common/src/test/java/org/opensearch/sdk/GetDataObjectRequestTests.java
@@ -9,23 +9,33 @@
 package org.opensearch.sdk;
 
 import org.junit.Before;
+import org.opensearch.search.fetch.subphase.FetchSourceContext;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
 
-public class DeleteDataObjectRequestTest {
+public class GetDataObjectRequestTests {
+
     private String testIndex;
     private String testId;
+    private FetchSourceContext testFetchSourceContext;
 
     @Before
     public void setUp() {
         testIndex = "test-index";
         testId = "test-id";
+        testFetchSourceContext = mock(FetchSourceContext.class);
     }
 
-    public void testDeleteDataObjectRequest() {
-        DeleteDataObjectRequest request = new DeleteDataObjectRequest.Builder().index(testIndex).id(testId).build();
+    public void testGetDataObjectRequest() {
+        GetDataObjectRequest request = new GetDataObjectRequest.Builder()
+            .index(testIndex)
+            .id(testId)
+            .fetchSourceContext(testFetchSourceContext)
+            .build();
 
         assertEquals(testIndex, request.index());
         assertEquals(testId, request.id());
+        assertEquals(testFetchSourceContext, request.fetchSourceContext());
     }
 }
