+++ common/src/test/java/org/opensearch/sdk/GetDataObjectResponseTests.java
@@ -9,24 +9,26 @@
 package org.opensearch.sdk;
 
 import org.junit.Before;
+import org.opensearch.core.xcontent.XContentParser;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
 
-public class PutDataObjectResponseTest {
+public class GetDataObjectResponseTests {
 
     private String testId;
-    private boolean testCreated;
+    private XContentParser testParser;
 
     @Before
     public void setUp() {
         testId = "test-id";
-        testCreated = true;
+        testParser = mock(XContentParser.class);
     }
 
-    public void testPutDataObjectResponse() {
-        PutDataObjectResponse response = new PutDataObjectResponse.Builder().id(testId).created(testCreated).build();
+    public void testGetDataObjectResponse() {
+        GetDataObjectResponse response = new GetDataObjectResponse.Builder().id(testId).parser(testParser).build();
 
         assertEquals(testId, response.id());
-        assertEquals(testCreated, response.created());
+        assertEquals(testParser, response.parser());
     }
 }
